+++ cpp/src/io/orc/writer_impl.hpp
@@ -42,6 +42,31 @@ namespace io {
 namespace detail {
 namespace orc {
 
+/**
+ * @brief Chunked writer state struct. Contains various pieces of information
+ *        needed that span the begin() / write() / end() call process.
+ */
+struct orc_chunked_state {
+  /// The writer to be used
+  std::unique_ptr<writer>             wp;
+  /// Cuda stream to be used
+  cudaStream_t                        stream;
+  /// Overall file metadata.  Filled in during the process and written during write_chunked_end()
+  cudf::io::orc::FileFooter           ff;
+  cudf::io::orc::Metadata             md;
+  /// current write position for rowgroups/chunks
+  size_t                              current_chunk_offset;
+  /// optional user metadata
+  table_metadata const*               user_metadata = nullptr;
+  /// only used in the write_chunked() case. copied from the (optionally) user supplied
+  /// argument to write_chunked_begin()
+  table_metadata_with_nullability     user_metadata_with_nullability;
+  /// special parameter only used by detail::write() to indicate that we are guaranteeing
+  /// a single table write.  this enables some internal optimizations.
+  bool                                single_write_mode = false;
+};
+
+
 // Forward internal classes
 class orc_column_view;
